+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix incorrect condition when deciding which snarked ledgers need to be synchronize during bootstrap which would result in a failure when bootstrapping the node during the second epoch.
+- Correctly handle the situation in which the best tip changes during staged ledger reconstruction, causing the reconstruct to produce a stale result.
 
 ## [0.3.1] - 2024-04-05